third_party/blink/web_tests/http/tests/fetch/serviceworker/thorough/scheme-blob-expected.txt

CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/serviceworker/resources/fetch-access-control-login.html' attempted to load subresource at 'https://username1s:password1s@127.0.0.1:8443/serviceworker/resources/fetch-access-control.php?Auth' over broken HTTPS.
CONSOLE ERROR: Mixed Content: The page at 'https://localhost:8443/serviceworker/resources/fetch-access-control-login.html' attempted to load subresource at 'https://username2s:password2s@localhost:8443/serviceworker/resources/fetch-access-control.php?Auth' over broken HTTPS.
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = 1 duplicate test name: "scheme-blob"
PASS Startup
PASS scheme-blob
PASS scheme-blob
Harness: the test ran to completion.
